// src/interfaces/cartitem.interface.ts
export type CartItemType = {
    /* id: number;
    category: string;
    description: string;
    image: string;
    price: number;
    title: string;
    amount: number; */

    id: number;
    name: string;
    status: string;
    species: string;
    type: string;
    gender: string;
    origin: object;
    location: object;
    image: string;
    episode: [];
    url: string;
    created: string;
    amount:number;
};
  